森林防火/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
SQLAlchemy==1.4.23
numpy==1.21.2
pandas==1.3.3
scikit-learn==1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
pymysql==1.0.2 